update to 2.2.2-BETA
It has a number of new features and performance improvements including
much-expanded header caching and support for the IMAP SASL-IR and BINARY
extensions.  This release also has a large number of bugfixes in the
Cyrus Murder support of virtdomain/altnamespace/unixhierachysep, as well
as a significant improvement in how Cyrus Master tracks processes.

Full changes are available in doc/changes.html in the distribution, sites
that are upgrading to this version will also want to read
doc/install-upgrade.html.