Finish update to e2fsprogs-1.35. The cvs was locked when I was working on this.
This also splits out the com_err(3) related files and uses the new
wip/com_err package.

This adds patches to make sure that libuuid's and com_err files
are not built and installed -- and to make sure the installed
package versions are used.

Added a E2FSPROGS_VERSION variable that can be used by other packages
using this distfile.

Moved the don't install fsck wrapper to the Makefile, since
it is just noise for the the Makefile.common.