This is the same unrealircd that I hacked at @6 months ago.  They have gone release now and I decided to give it another go a pkgsrc'ing it.
ipv6 support is busted so don't even bother but the rest of the
functionality looks ok.   The runtime setuid/setgid hack is mine so
if that breaks you can blame me.  It will probably sit here for a while
untill I get a last few issues sorted out with it.  I will also need to get
the developers to have a look at why ipv6 isn't working correctly.