Remove the xlibs meta-package. xlibs is getting out-of-date. It is not maintained. It was experimental.
The new official project is modular X.org. It is in release
candidate 0 and release candidate 1 is expected within a day.

Feel free to clean up any leftover xlibs work here in pkgsrc-wip
to use modular X.org instead. (Some packages may be updated
and in some cases, the distfile names have changed so new
packages can be created and old, no longer used packages removed.)

Also todo: make a meta package for modular-xorg that will be
equivalent to xorg-libs and can be used as an X11_TYPE.