Update to 1.1.2
15.2.2006

Oops, the last release has 2 problems:

The program was not compilable without SDL-Mixer installed

There was a problem in the makefiles that made it stop with an error when trying to install

Both problems are hotfixed in a file called toppler-1.1.2a. This is still the same release it just fixes the 2 problems. Only people downloading the source are affected.
13.2.2006

After a long break release 1.1.2 is ready for download. Changes include:

Another build system update (by Volker Grabsch).

Background music (by Angel Keha). Because the mp3 file is twice as large as the whole game I don't include it in the normal distribution. But you must download it separately. Just put it into the directory with the toppler data files and backround music should play

Two new loactisations: Portugese and Swedish. Localisations are not available for the windows binary, sorry.

Removed one rare possibility where the green guy could get stuck between tower elements
21.12.2004

Jason has provided a binary for the release 1.1.1. Download from the usual place
20.12.2004

There has been a bug in the windows version after all. So there is now also a 1.1.1 release of the windows binary.
18.12.2004

A small update to Toppler (1.1.1): Last release didn't really add the promised new mission to the source package. The new release also contains updates to the German and French translations and a completely new Finnish translation (Thanks go to Basse and Bill).

As the windows version doesn't use translations there is no need to update this package. You can still download the version 1.1.0 and get the right package.

There is now also an OSX version of the 1.1.0 release available. Please download it from here
13.12.2004

Finally there is a new version (1.1.0) of Tower Toppler. There are not many new features. A few smaller bugs have been squashed (the toppler could get caught, when an elevator came down and it was too near to it, one meny point wasn't saved)

The game features now joystick support

For leveldesigner it might be nice to know that they can now create elevators thas pass doors

Finally the windows problem with the strange .dll should be fixed

The version number change is because I merged the development and the "stable" version so beware there might be bugs that I have just not seen