Add editors/emacs/patches/patch-az. Commit message was:
revision 1.14
date: 2006/04/05 22:22:16;  author: markd;  state: Exp;  lines: +11 -169
Pass the correct sized argument to sbrk() when trying to reduce the break
so that on 64bit systems it is actually a negative number, not a very
large positive one.  Should fix PR pkg/29351.
Thanks to Martijn van Buul for giving me access to an amd64 box so I
could track this down.  Bump PKGREVISION.


Don't bump PKGREVISION here, because it's a daily CVS snapshot anyway.