This is xorg-server-1.3.99.0, the newest version of the Xorg server. It depends upon wip/pixman, which should install OK.
NOTE: all dependencies of this package also depend upon pixman, so make
sure to use the included buildlink3.mk file or otherwise buildlink
in these dependencies.  Also, you will need to use the GIT version
of xf86-input-keyboard to get the required header files.  I'll try to
have a package here of that soon.  I'm using this package now, and
it's quite stable--including DRI.

The dbus backend is untested.