Changes v3.1
2006/03/07	Its been out there, and the only bug reported in the 2 months
since the last beta release was a 3 line minor fix.
So 3.1 is the new official, stable release.
The list of user defined database paths were not being parsed correctly.
Thanks to Vincent Lefevre for finding the bug and fixing it.


Changes v3.0 beta r3:

2006/01/06	The environment variable LOCATE_PATH was not being used properly.
It will now be used the same as in v2.7.
slocate was returning non zero for -h, --help, -V and --version.
It now returns zero.
PRUNEFS was being ignored in /etc/updatedb.conf
Added documentation describing how /etc/updatedb.conf is used and what is
supported to the README and the slocate.1 and updatedb.1 man pages.


Changes v3.0 beta r2:

2006/01/05	Added fix to compensate for default unsigned chars on PPC.
Fixed segfault due to 0 length char array. oops.
Fixed a bug where dangling links were ignored.

Changes v3.0 beta r1:

2005/12/19	A complete rewrite. The idea being that when I first wrote this
I was still a bit of a newbie. This version has a better design
and will be easier to debug and maintain. I have also received too many
database corruption emails and I wanted to redesign the encode/decode routines
to make sure that this was fixed.
This version should have the same functionality of v2.7.
Free time is rare so I havn't had any to add any of the features
people have been asking for. Feature patches are welcome!