This commit updates tor to 0.2.1.18.
The only change, when compared to 0.2.1.18-rc
is releated to RPMs and therefore shouldn't
affect pkgsrc users.

The real difference is that this release is
tagged as stable and will move to net/tor.

If you want to track the stable branch please
switch to net/tor.

wip/tor-dev will soon be updated to an unstable
version of the 0.2.2.x branch.