Initial commit of FastBit, version 1.2.0.
FastBit is an open-source data processing library following the
spirit of NoSQL movement. It offers a set of searching functions
supported by compressed bitmap indexes. It treats user data in the
column-oriented fashion similar to well-known database management
systems such as Sybase IQ, MonetDB, and Vertica. It is designed to
accelerate user's data selection tasks without imposing undue
requirements. In particular, the user data is NOT required to be
under the control of FastBit software.