Import tap-plugins-0.7.2 as wip/tap-plugins.
TAP-plugins is short for Tom's Audio Processing plugins. It is a bunch of LADSPA
plugins for digital audio processing, intended for use in a professional DAW
environment such as Ardour. These plugins should compile and run on any recent
(that is, not seriously outdated) GNU/Linux system. They don't require any
special libraries besides the standard GNU C and math libraries, which are
expected to be provided on the machine used for compiling.