Update Rgraphviz to version 1.34.2. Changes unknown.
This upgrade uses Bioconductor 2.10, so as Rgraphviz 2.2.1 (latest)
comes with built-in graphviz sources and requires tweaking of
R_PKG_INSTALL_ARGS in order to use an external one.  Moreover, it
introduces grid::bezierPoints dependency which available only
starting from R 2.15, without noting this dependency on the
Bioconductor page, and pkgsrc still has 2.14.2 at the moment.