Import p5-Class-DBI-BaseDSN-1.22 as wip/p5-Class-DBI-BaseDSN.
Class::DBI::BaseDSN acts as a placeholder for a base class which will be
switched for a specific Class::DBI extension when you specify the dsn of the
database to connect to. For example in this case, the Class::DBI::BaseDSN will
replace itself with Class::DBI::mysql when the set_db call is executed.