Import evas-1.7.4 as wip/evas.
Evas is a clean display canvas API for several target display
systems that can draw anti-aliased text, smooth super and sub-sampled
scaled images, alpha-blend objects and much more.

It abstracts any need to know much about what the characteristics
of your display system are or what graphics calls are used to draw
them and how. It deals on an object level where all you do is create
and manipulate objects in a canvas, set their properties, and the
rest is done for you.

Evas optimises the rendering pipeline to minimise effort in redrawing
changes made to the canvas and so takes this work out of the
programmers hand, saving a lot of time and energy.

It's small and lean, designed to work on embedded systems all the
way to large and powerful multi-cpu workstations. It can be compiled
to only have the features you need for your target platform if you
so wish, thus keeping it small and lean. It has several display
back-ends, letting it display on several display systems, making
it portable for cross-device and cross-platform development.