A temp. fix to get node compiling on OS X under pkgsrc-wic.
Gyp does not know about the differences in libtool, so use the one it
expects (/usr/bin/libtool) if OS is Darwin.

OpenSSL is included in node as a static library.

If using node-modules that require OpenSSL (node-guardtime), then
node must be compiled with a shared openssl.  If the option 'openssl'
is specified, we will build node against a shared openssl and depending on
PREFER_PKGSRC, this will either be security/openssl or system openssl.