Update to 2.17.14:
2.17.14:
We are happy to announce the release of LilyPond 2.17.14. This
release contains the usual number of bugfixes (including putting
span bars back where they should be) and enhancements, and contains
some work in progress. You will have access to the very latest
features, but some may be incomplete, and you may encounter bugs
and crashes. If you require a stable version of Lilypond, we
recommend using the 2.16 version.