Add the latest development version of graphics/libcaca.
libcaca is a graphics library that outputs text instead of pixels,
so that it can work on older video cards or text terminals. It is
not unlike the famous AAlib library, with the following improvements:

 * Unicode support
 * 2048 available colours (some devices can only handle 16)
 * dithering of colour images
 * advanced text canvas operations (blitting, rotations)

Libcaca works in a text terminal (and should thus work on all Unix
systems including Mac OS X) using the S-Lang or ncurses libraries.
It also works natively on DOS and Windows.