Import py27-python-application-1.4.0 as wip/py-python-application.
This package is a collection of modules that are useful when building python
applications. Their purpose is to eliminate the need to divert resources into
implementing the small tasks that every application needs to do in order to run
successfully and focus instead on the application logic itself.

The modules that the application package provides are:

 * process - UNIX process and signal management;
 * python - python utility classes and functions;
 * configuration - a simple interface to handle configuration files;
 * log - an extensible system logger for console and syslog;
 * debug - memory troubleshooting and execution timing;
 * system - interaction with the underlying operating system;
 * notification - an application wide notification system;
 * version - manage version numbers for applications and packages;
 * dependency - verify package dependencies at runtime;