Patch from "John D. Baker" <jdbaker@mylinuxisp.com>:
Oddities with the runtime loader caused incompatible libraries with same
major version to be loaded.  Specifically, for a time, "ffmpeg" and
"ffmpeg010" both had "libavformat.so.51.x" and when asked for
"libavformat.so.51", the loader would attempt to load the one from
"ffmpeg" instead of "ffmpeg010".

It seems this can be gotten around at build time by the order in which
"buildlink3.mk" files are included--specifically, that for packages
requiring "ffmpeg010" over "ffmpeg", it should be first in the list.