Import openxcom-0.9nb1 as wip/openxcom.
OpenXcom is an open-source clone of the original UFO: Enemy Unknown
(X-Com: UFO Defense in USA), licensed under the GPL and written in
C++ / SDL.

The goal of the project is to bring back the tried and true feel
of the original with none of the issues. All the same graphics,
sound and gameplay with a brand new codebase written from scratch.

*** Note that this package provides a clone of the game's
*** engine only, it requires the original data files!