Import ossp-val-0.9.4 as wip/ossp-val.
OSSP val is a flexible name to value mapping library for C variables. It is a
companion library to OSSP var. It allows one to access C variables through name
strings, although the C language does neither provide such a dedicated facility
nor an evaluation construct (which could be used to implement such a facility
easily).

In general, this is used for accessing C variables without having to know the
actual symbol/address/reference. The typical use cases are in combination with
flexible configuration parsing and supporting loosly-coupled DSO-based module
architectures.