Import am-utils-6.2nb20130830 as wip/am-utils-git.
An automounter maintains a cache of mounted file systems. File systems
are mounted on demand when they are first referenced, and unmounted
after a period of inactivity. This helps to centralize all file system
access, provide a uniform site-wide namespace, and minimize downtimes
for clients.  The Berkeley Automounter, Amd, may be used as a
replacement for Sun's automounter.

This package builds a development version from the tip of the git
source repository.