Import slurm-2.6.4 as wip/slurm.
SLURM is an open-source resource manager designed for Linux
clusters of all sizes. It provides three key functions. First it
allocates exclusive and/or non-exclusive access to resources
(computer nodes) to users for some duration of time so they can
perform work. Second, it provides a framework for starting,
executing, and monitoring work (typically a parallel job) on a
set of allocated nodes. Finally, it arbitrates contention for
resources by managing a queue of pending work.

This is an update to v2.6.4 with additional prerequisites of hwloc
and hdf5.