Import py27-easydev-0.6.3 as wip/py-easydev.
The package easydev provides utilities that are of general usage for the
development of Python packages. For instance, it provides the sphinx templates
being used for this documentation. It is also used by other packages where
documentation using Sphinx is being used (e.g., rtools, spectrum). It provides
tools such as multisetup (to ease the development of several packages within a
single namespace), get_share_directory of any package, functions that are used
often such as type checking.