Import solarus-zsdx-1.7.1 as wip/solarus-zsdx.
The Legend of Zelda: Mystery of Solarus DX is set to be a direct sequel to The
Legend of Zelda: A Link to the Past on the SNES, using the same graphics and
game mechanisms. Zelda Mystery of Solarus DX is the first game made with the
Solarus engine and in fact, Solarus was primarily designed for this game.

Zelda: Mystery of Solarus DX is actually an enhanced remake of a first creation,
Zelda: Mystery of Solarus. This first creation, developed with RPG Maker 2000,
was released in 2002 and was only available in French. The DX project was
unveiled on April 1st, 2008. Its objectives are to correct the many flaws of its
predecessor: the battle system, the bosses, the use of items, etc.