Bump 2.0.1RC3 to RC4 -------------------- This release has the following corrections since RC3:
1. Reverted the -RC3 change that converted the Canadian provinces from "PQ" to "QC".
This information is managed internally by LoTW in a form which is not ADIF. Making this
change would have rendered station locations in Quebec and Newfoundland unusable.
LoTW will need to make a correction so that when it exports logs, it exports these as the
ADIF standard "QC" and "NL".

2. Correct a defect in the way the start-up threads were being managed that caused TQSL to
hang on some UNIX variants.

3. Change the way pending callsign requests were being processed to avoid using a function
that fails on OpenSuSE Linux.

4. When signing a log, refuse to use a station location that does not have a valid callsign
certificate associated with the location.

5. Correct a defect that would cause TQSL to hang when certain command line arguments used
empty strings (-p, for example).

6. Update the messages displayed during callsign certificate creation to clarify that the
user should wait for a certificate reply from LoTW, to better explain when signed/unsigned
certificate requests should be used, and to not prompt for a signing certificate if none is
available.