Update to 2.0.1-RC7 ----------------------- An updated release candidate for TQSL 2.0.1 is now available. It corrects two serious defects in RC6 that made it unacceptable for public release.
1. When a callsign certificate is about to expire (within 60 days), TQSL prompts the user if
they want to renew that certificate. This prompt happens even if the user has already
submitted a renewal. If the user clicked "No" to not renew again, TQSL could get stuck in a
loop displaying "Invalid argument" repeatedly.

2. On Linux systems, using the preferences pane to change the directory for automatic backups
behaved very badly (not accepting user selections), and could crash TQSL. As this is a
component of the operating system that TQSL cannot correct, the directory browser is no
longer used on Linux systems (it performs as expected in Windows and MacOS.) For Linux, the
user is presented with a text box for entering the directory name for the backups.

In addition, this release will search for files with .ADI or .ADIF extensions when opening
log files.