Import py27-pdfposter-0.6.0 as wip/py-pdfposter.
Pdfposter can be used to create a large poster by building it from
multiple pages and/or printing it on large media. It expects as
input a PDF file, normally printing on a single page. The output
is again a PDF file, maybe containing multiple pages together
building the poster. The input page will be scaled to obtain the
desired size.

This is much like poster does for Postscript files, but working
with PDF. Since sometimes poster does not like your files converted
from PDF. :-) Indeed pdfposter was inspired by poster.