Import webkit-gtk-2.4.1 as wip/webkit-gtk.
WebKit is an open source web browser engine. WebKit is also the name of
the Mac OS X system framework version of the engine that's used by
Safari, Dashboard, Mail, and many other OS X applications. WebKit's HTML
and JavaScript code began as a branch of the KHTML and KJS libraries
from KDE.

This is the GTK2+ port of the engine.

Please note that this package is intended to update www/webkit-gtk but it is
still *not* buildable. Please read TODO and help to update it! ;)