Bump 2.0.2RC3 to 2.0.2RC4. The letter from developer: ------------------- Things have been pretty quiet for 2.0.2, so I've posted a kit with the final 2.0.2 changes.
Two significant changes are included in -RC4:

1. When prompting a user for the password to unlock their
callsign certificate while signing a log, don't pop up the
main TQSL window if TQSL is being run in batch mode (usually
this is when a logging program is using TQSL to sign).

2. If the QSO database can't be opened and can't be repaired,
delete it and start over.

As usual, please let me know if you see any problems. If
nothing is seen, I'll be posting an announcement for these on
ARRL-LOTW for final testing next weekend.