(Authors note): As I've had no comments on the 2.0.2-rc5 beta, I've decided that it's time to release it.
The kits on sourceforge have been updated to the final v2.0.2 release.

https://sourceforge.net/projects/trustedqsl/files/TrustedQSL/v2.0.2/

The README there is the preliminary release notes.
Thanks to everyone for the help with testing and reporting issues.

(pkgsrc)
README (as is) doesn't look for regular release notes, but usage of
the software. (So it is not included yet as a binary or commit log.)