Half-baked postgis2 package.
I've had this lying around for a long time.  I'm putting it in wip so
that others can work on it as well.  The notion is that it enter
pkgsrc as postresql-postgis2 so that anyone still using postgis 1.x
can migrate, and because it's likely some future version will need to
be parallel-buildable (even if not parallel-installable).