Import tuxpaint-cvs-2014.08.17nb20140817 as wip/tuxpaint-cvs.
Tux Paint is a free drawing program designed for young children
(kids ages 3 and up). It has a simple, easy-to-use interface, fun
sound effects, and an encouraging cartoon mascot who helps guide
children as they use the program.

It provides a blank canvas and a variety of drawing tools to help
your child be creative.

This package contains the latest CVS version of tuxpaint.