Import py-qt5-5.2.1 as wip/py-qt5. Packages fine, but needs more testing, e.g. a working package that uses it.
PyQt is a set of Python bindings for the Qt toolkit. PyQt brings
together the Qt C++ cross-platform toolkit and the cross-platform
interpreted language Python.  PyQt contains 300 classes and over
5,750 functions and methods.

PyQt also includes the pyuic and pylupdate utilities which correspond
to the Qt uic and lupdate utilities.  pyuic converts the GUI designs
created with Qt Designer to executable Python code.  pylupdate
scans Python code, extracts all strings that are candidates for
internationalisation, and creates an XML file for use by Qt Linguist.

This package is for QT5.