Import gnuradio-zeromq-3.7.5 as wip/gnuradio-zeromq.
GNU Radio is a collection of software that when combined with minimal hardware,
allows the construction of radios where the actual waveforms transmitted and
received are defined by software. What this means is that it turns the digital
modulation schemes used in today's high performance wireless devices into
software problems.

This contains all of the zeromq (net/zeromq, 0MQ lightweight messaging
kernel) blocks, utilities, and examples.