Import fastjar-0.94 as wip/fastjar.
Fastjar is a version of Sun's 'jar' utility, written entirely in C,
and therefore quite a bit faster.  Fastjar can be up to 100x faster
than the stock 'jar' program running without a JIT.  Currently,
the author is working on adding all the features present in the sun
utility.  At the moment, implemented features are:

	* Archive creation
	* Verbose/quiet output
	* stdout vs. file output
	* Manifest file support
	* Deflation or storage
	* Changing to a dir and adding files (-C)
	* Archive content listing (-t)
	* Archive extraction (-x)