Import fuse-mp3fs-0.91 as wip/fuse-mp3fs. This is the version that was in filesystems/fuse-mp3fs for a short while with two additional patches.
MP3FS is a read-only FUSE filesystem which transcodes audio formats
(currently FLAC) to MP3 on the fly when opened and read. This was
written to enable using a FLAC collection with software and/or
hardware which only understands MP3.

It is also a novel alternative to traditional MP3 encoder applications.
Just use your favorite file browser to select the files you want
encoded and copy them somewhere!