Import felt-3.07 as wip/felt.
FElt is a free system for introductory level finite element
analysis. It is primarily intended as a teaching tool for introductory
type courses in finite elements - probably in the
mechanical/structural/civil fields. In a command line environment,
FElt uses an intuitive, straightforward input syntax to describe
problems. It also includes a graphical user interface for workstations
that allows the user to set-up, solve and post-process the problem in
a single CAD-like environment.

From the end-user point of view, FElt consists of six programs: felt,
burlap, velvet, corduroy, patchwork, and yardstick. felt is the basic
command-line application; burlap is an interactive Matlab-like
environment for scripting your own elements and analyses; velvet is
the primary GUI interface into the bulk of the functionality in FElt;
corduroy is command-line program for automatic element generation;
patchwork is a command-line application for file format conversion to
and from the FElt syntax; and yardstick is a simple program for
problem scaling and unit conversion.