Add wip/rpm5
RPM is a powerful and mature command-line driven package management system
capable of installing, uninstalling, verifying, querying, and updating Unix
software packages. Each software package consists of an archive of files
along with information about the package like its version, a description,
and the like. There is also a library API, permitting advanced developers
to manage such transactions from programming languages such as C, Perl or
Python.

This package was started but never finished. As an abandoned piece of work
better to keep it her then put to trash.