Import mousetweaks-3.12.0 as wip/mousetweaks.
The Mousetweaks package provides mouse accessibility enhancements
for the GNOME desktop. These enhancements are:

1. It offers a way to perform the various clicks without using any
hardware button.

2. It allows users to perform a right click by doing a click&hold
of the left mousebutton. (For a left-handed mouse user, the termes
left and right have to be inverted.)

3. It provides an applet that the user can install on a panel. This
applet creates an area on the panel into which the pointer can be
captured until the user releases it with a predefined button and
modifier combination.

The options can be accessed through the Accessibility tab of the
Mouse Preferences of GNOME Control Center or through command-line.