Import initial work of ltp (Linux Test Project) as wip/ltp v. git
The Linux Test Project is a joint project started by SGI, developed and
maintained by IBM, Cisco, Fujitsu, SUSE, Red Hat and others, that has a goal to
deliver test suites to the open source community that validate the reliability,
robustness, and stability of Linux.

The LTP testsuite contains a collection of tools for testing the Linux kernel
and related features.