Import cvs-fast-export-1.29 as wip/cvs-fast-export.
Does not compile on NetBSD-7.99.4/amd64 due to:
revcvs.c:80:43: error: 'PTHREAD_RECURSIVE_MUTEX_INITIALIZER_NP' undeclared here (not in a function)
 static pthread_mutex_t dir_bucket_mutex = PTHREAD_RECURSIVE_MUTEX_INITIALIZER_NP;
                                           ^

Perhaps someone has an idea.

This program analyzes a collection of RCS files in a CVS
repository (or outside of one) and, when possible, emits an
equivalent history in the form of a fast-import stream. Not all
possible histories can be rendered this way; the program tries to
emit useful warnings when it can't. The program can also produce
a visualization of the resulting commit DAG in the DOT format
handled by the graphviz suite.