Import netsurf-all-3.2 as wip/netsurf.
Small as a mouse, fast as a cheetah and available for free.
NetSurf is a web browser for RISC OS and UNIX-like platforms.

Whether you want to check your webmail, read the news or post to discussion
forums, NetSurf is your lightweight gateway to the world wide web.
Actively developed, NetSurf is continually evolving and improving.

XXX: please give a look to TODO, there is a problem with some https:// sites and
certificates. Apart that it seems that it works correctly.