Import kibana-4.0.0 as wip/kibana.
Functional... rc.subr script is hackish to manage the pid. Not sure if
there's a better way to handle that.

Node.js build env probably doesn't play well with bulk builds etc at this
point. It appears to install and modify things outside the pkgsrc env.

Kibana is an open source (Apache Licensed), browser based analytics and
search dashboard for Elasticsearch. Kibana is a snap to setup and start
using.
Kibana strives to be easy to get started with, while also being flexible
and powerful, just like Elasticsearch.