Import /usr/pkgsrc-1/wip/ganglia-monitor-core ganglia-monitor-core-3.6.0 as wip/ganglia-monitor-core.
Ganglia provides a complete real-time monitoring and execution
environment that is in use by hundreds of universities, private and
government laboratories and commercial cluster implementors around the
world. Whether you want to monitor hundreds of computers in real-time
across a university campus or around the world, ganglia is for you.

WWW: http://ganglia.info/