Import kwindowsystem-5.9.0 as wip/kwindowsystem
Convenience access to certain properties and features of the windowing system.

KWindowSystem provides information about the windowing system and allows
interaction with the windowing system. It provides an high level API which is
windowing system independent and has platform specific implementations. This
API is inspired by X11 and thus not all functionality is available on all
windowing systems.

In addition to the high level API, this framework also provides several more
low level classes for interaction with the X Windowing System.