Update to 2.13.0 ---------------- In this release we have focused on improving performance and making Piwik much faster at loading and archiving reports. The dashboard and All websites dashboard will load faster than ever before! This is especially visible if you measure many websites (hundreds of thousands) within your Piwik, the improved Websites Manager will now let you search for websites and page through the list, and the ‘All Websites’ dashboard will now load correctly.
Performance was not our only focus and we also improved the
usability and design of some parts of Piwik (such as the
Updater and maintenance mode, with more design updates coming
in the next releases). Log Analytics, a very popular tool
within the Piwik community, has received several
improvements. Many other small bugs were closed, and a XSS
security bug was reported and fixed in this release.

97 tickets have been closed by 8 contributors!