Import pristine x11/gdm as wip/gdm
This package will be upgraded for Gnome 3

GDM is the GNOME Display Manager, it is the little proggie that runs
in the background, runs your X sessions, presents you with a login box
and then tells you to piss off because you forgot your password. It
does pretty much everything that you would want to use xdm for, but
doesn't involve as much crack.  It doesn't use any code from xdm, and
has a more paranoid and safer design overall.  It also includes many
features over xdm, the biggest one of which is that it is more user
friendly, even if your X setup is failing.  The goal is that users
should never, ever have to use the command line to customize or
troubleshoot gdm.  It of course supports xdmcp, and in fact extends
xdmcp a little bit in places where I thought xdm was lacking (but is
still compatible with xdm's xdmcp).