Import /home/bacon/Pkgsrc/pkgsrc-2/wip/fbsd2pkg fbsd2pkg-0.1 as wip/fbsd2pkg.
The fbsd2pkg command converts a FreeBSD port framework to a pkgsrc package.
Most common constructs are converted, but in most cases, manual editing is
required to reach a working package.

Fbsd2pkg is an alternative to port2pkg that tries to do more automatic
conversion as well as preserve all unconverted FreeBSD port syntax as
comments in the pkgsrc Makefile.

It is written in Bourne shell and Awk to avoid the need for any large
dependencies.